Implemented 3D domain splitting
Domain is decomposed such that communication cost is reduced. If
two decompositions have same communication cost, then the one with
lowest maximum load is picked. If both communication costs and
maximum loads are same, then the one with lowest load factor is
picked. If everything is same, then the first decomposition is
selected.